# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "BottleCRM API",
    "version": "1.0.0",
    "description": "Multi-tenant CRM API with JWT authentication",
    "pluginName": "BottleCRM智能助手",
    "pluginDesc": "BottleCRM客户关系管理系统AI助手，支持查询客户数据、创建线索任务、分析销售业绩等功能",
    "contact": "support@bottlecrm.com",
    "x-huawei-plugin-name": "BottleCRM智能助手",
    "x-huawei-plugin-desc": "BottleCRM客户关系管理系统AI助手，支持查询客户数据、创建线索任务、分析销售业绩等功能",
    "x-huawei-version": "1.0.0",
    "x-huawei-contact": "support@bottlecrm.com"
  },
...
  "servers": [
    {
      "url": "http://localhost:3002",
      "description": "Development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/tasks": {
      "get": {
        "summary": "Get all tasks for organization",
        "tags": [
          "Tasks"
        ],
        "parameters": [
          {
            "in": "header",
            "name": "X-Organization-ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "status",
            "schema": {
              "type": "string"
            },
            "description": "Filter tasks by status"
          },
          {
            "in": "query",
            "name": "priority",
            "schema": {
              "type": "string"
            },
            "description": "Filter tasks by priority"
          },
          {
            "in": "query",
            "name": "ownerId",
            "schema": {
              "type": "string"
            },
            "description": "Filter tasks by owner ID"
          },
          {
            "in": "query",
            "name": "accountId",
            "schema": {
              "type": "string"
            },
            "description": "Filter tasks by account ID"
          },
          {
            "in": "query",
            "name": "contactId",
            "schema": {
              "type": "string"
            },
            "description": "Filter tasks by contact ID"
          },
          {
            "in": "query",
            "name": "leadId",
            "schema": {
              "type": "string"
            },
            "description": "Filter tasks by lead ID"
          },
          {
            "in": "query",
            "name": "opportunityId",
            "schema": {
              "type": "string"
            },
            "description": "Filter tasks by opportunity ID"
          },
          {
            "in": "query",
            "name": "caseId",
            "schema": {
              "type": "string"
            },
            "description": "Filter tasks by case ID"
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "default": 50
            },
            "description": "Limit number of results"
          },
          {
            "in": "query",
            "name": "offset",
            "schema": {
              "type": "integer",
              "default": 0
            },
            "description": "Offset for pagination"
          }
        ],
        "responses": {
          "200": {
            "description": "List of tasks"
          }
        }
      },
      "post": {
        "summary": "Create a new task",
        "tags": [
          "Tasks"
        ],
        "parameters": [
          {
            "in": "header",
            "name": "X-Organization-ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "subject"
                ],
                "properties": {
                  "subject": {
                    "type": "string"
                  },
                  "status": {
                    "type": "string",
                    "enum": [
                      "Not Started",
                      "In Progress",
                      "Completed",
                      "Deferred",
                      "Waiting"
                    ],
                    "default": "Not Started"
                  },
                  "priority": {
                    "type": "string",
                    "enum": [
                      "High",
                      "Normal",
                      "Low"
                    ],
                    "default": "Normal"
                  },
                  "dueDate": {
                    "type": "string",
                    "format": "date-time"
                  },
                  "description": {
                    "type": "string"
                  },
                  "ownerId": {
                    "type": "string",
                    "description": "UUID of the user who owns this task"
                  },
                  "accountId": {
                    "type": "string",
                    "description": "UUID of the related account"
                  },
                  "contactId": {
                    "type": "string",
                    "description": "UUID of the related contact"
                  },
                  "leadId": {
                    "type": "string",
                    "description": "UUID of the related lead"
                  },
                  "opportunityId": {
                    "type": "string",
                    "description": "UUID of the related opportunity"
                  },
                  "caseId": {
                    "type": "string",
                    "description": "UUID of the related case"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Task created successfully"
          },
          "400": {
            "description": "Validation error"
          }
        }
      }
    },
...
    "/tasks/{id}/comments": {
      "post": {
        "summary": "Add a comment to a task",
        "tags": [
          "Tasks"
        ],
        "parameters": [
          {
            "in": "header",
            "name": "X-Organization-ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Task ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "body"
                ],
                "properties": {
                  "body": {
                    "type": "string"
                  },
                  "isPrivate": {
                    "type": "boolean",
                    "default": false
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Comment added successfully"
          },
          "404": {
            "description": "Task not found"
          }
        }
      }
    }
  },
...
}